# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/role": {
      "patch": {
        "operationId": "UserController_updateUserRole",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Academic Management System API",
    "description": "In-memory NestJS backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateRoleInputDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "student",
              "faculty",
              "admin",
              "academic_head"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "UserOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
...
    }
  }
}